fileshare/requirements.txt
aiohttp
aiofiles
python-dotenv
asyncpg
pyrofork
tgcrypto
aiohttp
asyncpg
pyrogram
TgCrypto
